+++ faq-bundle/CHANGELOG.md
@@ -1,3 +1,5 @@
 # Contao FAQ bundle change log
 
 ### DEV
+
+ * Support sorting enclosures (see contao/calendar-bundle#16).